#19 Refactor : gateway -  Utils 분리
